seperate public and protected claim routes
